+++ packages/config/__tests__/ProfileInfo.TeamConfig.test.ts
@@ -753,25 +753,25 @@ describe("TeamConfig ProfileInfo tests", () => {
             }
         });
 
-        it("should find correct offSchema args", async () => {
+        it("should find correct inSchema args", async () => {
             const profInfo = createNewProfInfo(teamProjDir);
             await profInfo.readProfilesFromDisk();
             const profAttrs = profInfo.getAllProfiles()[7] as IProfAttrs;
             const mergedArgs = profInfo.mergeArgsForProfile(profAttrs);
 
             const expectedArgs = [
-                { argName: "host", offSchema: false },
-                { argName: "port", offSchema: false },
-                { argName: "responseFormatHeader", offSchema: true },
-                { argName: "fakeOffSchemaArg", offSchema: true },
-                { argName: "user", offSchema: false },
-                { argName: "password", offSchema: false },
-                { argName: "rejectUnauthorized", offSchema: false },
+                { argName: "host", inSchema: true },
+                { argName: "port", inSchema: true },
+                { argName: "responseFormatHeader", inSchema: false },
+                { argName: "fakeOffSchemaArg", inSchema: false },
+                { argName: "user", inSchema: true },
+                { argName: "password", inSchema: true },
+                { argName: "rejectUnauthorized", inSchema: true },
             ];
 
             expect(mergedArgs.knownArgs.length).toBe(expectedArgs.length);
             for (const [idx, arg] of mergedArgs.knownArgs.entries()) {
-                expect(arg.offSchema).toEqual(expectedArgs[idx].offSchema);
+                expect(arg.inSchema).toEqual(expectedArgs[idx].inSchema);
             }
         });